adds colors file, reorganized files into folders
