+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 
 import Header from './components/Header';
-import StartGameScreen from './components/StartGameScreen';
+import StartGameScreen from './screens/StartGameScreen';
 
 export default function App() {
   return (
