+++ constants/colors.js
@@ -1,0 +1,4 @@
+export default {
+  primary: '#f7287b',
+  accent: '#c717fc',
+}
