+++ screens/StartGameScreen.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { View, Text, TextInput, StyleSheet, Button } from 'react-native';
 
-import Card from './Card';
+import Card from '../components/Card';
+import Colors from '../constants/colors';
 
 const StartGameScreen = props => {
   return (
@@ -11,8 +12,12 @@ const StartGameScreen = props => {
         <Text>Select a Number</Text>
         <TextInput />
         <View style={styles.buttonContainer}>
-          <Button title='Reset' />
-          <Button title='Confirm' />
+          <View style={styles.button}>
+            <Button title='Reset' color={Colors.accent} />
+          </View>
+          <View style={styles.button}>
+            <Button title='Confirm' color={Colors.primary} />
+          </View>
         </View>
       </Card>
     </View>
@@ -39,6 +44,9 @@ const styles = StyleSheet.create({
     width: '100%',
     justifyContent: 'space-between',
     paddingHorizontal: 15
+  },
+  button: {
+    width: 100
   }
 });
